+++ backend/src/main/java/org/scoula/report/domain/ReportDTO.java
@@ -20,7 +20,7 @@ public class ReportDTO {
     private int totalScore; // 총점
     private Integer jeonseRate; // 전세가율
     private Boolean accordOwner; // 소유자, 계약자 일치 여부
-    private Integer maximumOfBond; // 채권최고액
+    private Long maximumOfBond; // 채권최고액
     private String useType; // 주용도
     private Boolean violationStructure; // 위반건축물 여부
     private Date contractStartDate; // 계약 시작일
