+++ backend/src/main/java/org/scoula/safety_inspection/infra/cors/dto/CopyOfRegisterDto.java
@@ -12,9 +12,9 @@ public class CopyOfRegisterDto {
     private String ownership;
     private String commonOwner;
     private Integer changeOwnerCount;
-    private Integer maximumOfBond;
+    private Long maximumOfBond;
 
-    public CopyOfRegisterDto(Integer analysisNo,Double ownerState, String ownership, String commonOwner, Integer changeOwnerCount, Integer maximumOfBond) {
+    public CopyOfRegisterDto(Integer analysisNo,Double ownerState, String ownership, String commonOwner, Integer changeOwnerCount, Long maximumOfBond) {
         this.analysisNo = analysisNo;
         this.ownerState = ownerState;
         this.ownership = ownership;
